+++ packages/video_player/video_player_platform_interface/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 6.0.3
 
-* Add the `VideoEventType.isPlayingStateUpdate` event to track changes in play / pause state with
+* Adds the `VideoEventType.isPlayingStateUpdate` event to track changes in play / pause state with
   the underlying video player.
 
 ## 6.0.2